Fixed LSP errors and formatted (flask runs)

+++ app/controllers/group_controller.py
@@ -2,14 +2,17 @@ from flask import Blueprint, abort, jsonify, request
 from app.services.group_service import GroupService
 
 
-def create_group_bp(group_service: GroupService):
+def create_group_bp(group_service: GroupService) -> Blueprint:
     group_bp = Blueprint("group", __name__)
 
     @group_bp.route("create", methods=["POST"])
     def create():
         try:
+            if request.json == None:
+                abort(400, "no body provided")
             group_name = request.json["group_name"]
-            group_service.create_group(group_name)
+            user_id = request.json["user_id"]
+            group_service.create_group(group_name, user_id)
             return jsonify({"msg": "Group created successfully"}), 201
         except ValueError as ve:
             abort(400, str(ve))
@@ -20,6 +23,8 @@ def create_group_bp(group_service: GroupService):
     @group_bp.route("join", methods=["POST"])
     def join():
         try:
+            if request.json == None:
+                abort(400, "no body provided")
             user_id = request.json["user_id"]
             group_id = request.json["group_id"]
             group_service.add_user_to_group(user_id, group_id)
@@ -32,9 +37,11 @@ def create_group_bp(group_service: GroupService):
     @group_bp.route("users", methods=["GET"])
     def users():
         try:
+            if request.json == None:
+                abort(400, "no body provided")
             group_id = request.args.get("group_id")
             group = group_service.find_group_by_id(group_id)
-            return jsonify(group["users"]), 200
+            return jsonify(group.users), 200
         except ValueError as ve:
             abort(400, str(ve))
         except Exception as e:
@@ -43,6 +50,8 @@ def create_group_bp(group_service: GroupService):
     @group_bp.route("leave", methods=["POST"])
     def leave_group():
         try:
+            if request.json == None:
+                abort(400, "no body provided")
             user_id = request.json["user_id"]
             group_id = request.json["group_id"]
             group_service.remove_user_from_group(user_id, group_id)
@@ -64,3 +73,6 @@ def create_group_bp(group_service: GroupService):
         #     abort(400, str(ve))
         # except Exception as e:
         #     abort(500, "Internal Server Error")
+
+
+    return group_bp
